refactor: rename LogoutButton to SignOutButton
