+++ packages/admin/cms-admin/src/index.ts
@@ -108,11 +108,11 @@ export { DropdownMenuItem } from "./common/DropdownMenuItem";
 export { useBuildInformation } from "./common/header/about/build-information/useBuildInformation";
 export { Header } from "./common/header/Header";
 export { UserHeaderItem } from "./common/header/UserHeaderItem";
-export { LogoutButton } from "./common/logoutButton/LogoutButton";
+export { SignOutButton } from "./common/signOutButton/SignOutButton";
 export type { TextMatch } from "./common/MarkedMatches";
 export { MarkedMatches } from "./common/MarkedMatches";
 export { findTextMatches } from "./common/MarkedMatches";
-export type { MasterMenuData, MasterMenuProps } from "./common/MasterMenu";
+export type { MasterMenuData, MasterMenuProps } from "./common/MasterMenu"; 
 export { MasterMenu, useMenuFromMasterMenuData } from "./common/MasterMenu";
 export type { MasterMenuRoutesProps } from "./common/MasterMenuRoutes";
 export { MasterMenuRoutes, useRoutePropsFromMasterMenuData } from "./common/MasterMenuRoutes";
